feat(socketio/parser): create parser mod to handle multiple parser
