+++ socketioxide/Cargo.toml
@@ -25,6 +25,7 @@ tower.workspace = true
 http.workspace = true
 http-body.workspace = true
 thiserror.workspace = true
+smallvec.workspace = true
 itoa.workspace = true
 hyper.workspace = true
 matchit.workspace = true
