+++ socketioxide/src/lib.rs
@@ -294,6 +294,7 @@ pub mod handler;
 pub mod layer;
 pub mod operators;
 pub mod packet;
+pub mod parser;
 pub mod service;
 pub mod socket;
 
@@ -307,7 +308,6 @@ mod client;
 mod errors;
 mod io;
 mod ns;
-mod parser;
 
 /// Socket.IO protocol version.
 /// It is accessible with the [`Socket::protocol`](socket::Socket) method or as an extractor
